streaming: fix Tx packetsToBatch clamp truncated through uint8_t
TxSetup clamped mTx.packetsToBatch (a uint32_t) with an explicit
std::clamp<uint8_t> template argument, narrowing both the value and the
upper bound dmaBufferSize/packetSize modulo 256. Batch counts above 255
were silently truncated, and when dmaBufferSize/packetSize was a multiple
of 256 the upper bound wrapped to 0, violating std::clamp's lo<=hi
precondition and yielding packetsToBatch==0 -> divide-by-zero in the FIFO
sizing. Use uint32_t to match the equivalent Rx path.

+++ src/streaming/TRXLooper.cpp
@@ -1101,7 +1101,7 @@ OpStatus TRXLooper::TxSetup()
     const auto dmaChunks{ mTxArgs.dma->GetBuffers() };
     const auto dmaBufferSize = dmaChunks.front().size;
 
-    mTx.packetsToBatch = std::clamp<uint8_t>(mTx.packetsToBatch, 1, dmaBufferSize / packetSize);
+    mTx.packetsToBatch = std::clamp<uint32_t>(mTx.packetsToBatch, 1u, dmaBufferSize / packetSize);
 
     std::vector<uint8_t*> dmaBuffers(dmaChunks.size());
     for (uint32_t i = 0; i < dmaChunks.size(); ++i)
